requirements.txt
asgiref==3.5.2
Django==4.1.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
Pillow==9.3.0
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.7
